Defined new traits that provide methods to aide in the implementation of php unit tests.
```
src/traits/PHPUnitConfigurationTests.php
src/traits/PHPUnitRandomValues.php
src/traits/PHPUnitTestMessages.php

```

Refactored `tests/PHPUnitTestUtilitiesTest.php` to use the new traits
which provide methods that replace the methods it originally defined.

```
tests/PHPUnitTestUtilitiesTest.php

```
