+++ src/traits/PHPUnitConfigurationTests.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Darling\PHPUnitTestUtilities\traits;
+
+trait PHPUnitConfigurationTests
+{
+
+    /**
+     * Test that PHPUnit tests run.
+     *
+     * If this test does not run then PHPUnit is not set up correctly.
+     *
+     */
+    public function test_php_unit_tests_are_run(): void
+    {
+        $this->assertTrue(
+            true,
+            $this->testFailedMessage(
+                $this,
+                'test_php_unit_tests_are_run',
+                'run if PHPUnit is set up correctly'
+            )
+        );
+    }
+
+}
